"Refactor Board and Inquery components to use React Query for fetching tarot cards and reading, and update get_tarot_reading function to return chat completion object"
